Add user when creating post and permit user when deleting

+++ app/Http/Controllers/PostController.php
@@ -37,7 +37,7 @@ class PostController extends Controller
         $params['category_id'] = request('category');
 
         // create new post
-        $post = Post::create($params);
+        $post = auth()->user()->posts()->create($params);
         $post->tags()->attach(request('tags'));
 
         session()->flash('success', 'The post was created');
@@ -70,9 +70,17 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        $post->tags()->detach();
-        $post->delete();
-        session()->flash('success', 'The post was deleted');
-        return redirect('posts');
+        if (auth()->user()->is($post->author))
+        {
+            $post->tags()->detach();
+            $post->delete();
+            session()->flash('success', 'The post was deleted');
+            return redirect('posts');
+        }
+        else
+        {
+            session()->flash('error', 'It was not your post');
+            return redirect('posts');
+        }
     }
 }
